Grape alfa test version (RUS)
Soon there will be an English build and everyone can try it. I made the first version of the assistant smd for beginners in cmd. This is my first build of this program, and so far there is nothing special about it, but this is the first step to the future assistant. if you are Russian and want to test it, then I allow you.